// client-app/src/api/generated.ts
export class Candidate implements ICandidate {
    id?: string;
    name?: string | undefined;
    skills?: string[] | undefined;
    profileUrl?: string | undefined;
    summary?: string | undefined;
    location?: Location;
    availableFrom?: Date;
    isSpecial?: boolean;

    constructor(data?: ICandidate) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            if (Array.isArray(_data["skills"])) {
                this.skills = [] as any;
                for (let item of _data["skills"])
                    this.skills!.push(item);
            }
            this.profileUrl = _data["profileUrl"];
            this.summary = _data["summary"];
            this.location = _data["location"] ? Location.fromJS(_data["location"]) : <any>undefined;
            this.availableFrom = _data["availableFrom"] ? new Date(_data["availableFrom"].toString()) : <any>undefined;
            this.isSpecial = _data["isSpecial"];
        }
    }

    static fromJS(data: any): Candidate {
        data = typeof data === 'object' ? data : {};
        let result = new Candidate();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        if (Array.isArray(this.skills)) {
            data["skills"] = [];
            for (let item of this.skills)
                data["skills"].push(item);
        }
        data["profileUrl"] = this.profileUrl;
        data["summary"] = this.summary;
        data["location"] = this.location ? this.location.toJSON() : <any>undefined;
        data["availableFrom"] = this.availableFrom ? this.availableFrom.toISOString() : <any>undefined;
        data["isSpecial"] = this.isSpecial;
        return data;
    }
}

export interface ICandidate {
    id?: string;
    name?: string | undefined;
    skills?: string[] | undefined;
    profileUrl?: string | undefined;
    summary?: string | undefined;
    location?: Location;
    availableFrom?: Date;
    isSpecial?: boolean;
}

export class Job implements IJob {
    id?: string;
    website?: string | undefined;
    title?: string | undefined;
    company?: string | undefined;
    location?: Location;
    language?: string | undefined;
    description?: string | undefined;
    salary?: Salary;
    posterName?: string | undefined;
    postedDate?: Date;
    url?: string | undefined;
    tags?: string[] | undefined;

    constructor(data?: IJob) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.website = _data["website"];
            this.title = _data["title"];
            this.company = _data["company"];
            this.location = _data["location"] ? Location.fromJS(_data["location"]) : <any>undefined;
            this.language = _data["language"];
            this.description = _data["description"];
            this.salary = _data["salary"] ? Salary.fromJS(_data["salary"]) : <any>undefined;
            this.posterName = _data["posterName"];
            this.postedDate = _data["postedDate"] ? new Date(_data["postedDate"].toString()) : <any>undefined;
            this.url = _data["url"];
            if (Array.isArray(_data["tags"])) {
                this.tags = [] as any;
                for (let item of _data["tags"])
                    this.tags!.push(item);
            }
        }
    }

    static fromJS(data: any): Job {
        data = typeof data === 'object' ? data : {};
        let result = new Job();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["website"] = this.website;
        data["title"] = this.title;
        data["company"] = this.company;
        data["location"] = this.location ? this.location.toJSON() : <any>undefined;
        data["language"] = this.language;
        data["description"] = this.description;
        data["salary"] = this.salary ? this.salary.toJSON() : <any>undefined;
        data["posterName"] = this.posterName;
        data["postedDate"] = this.postedDate ? this.postedDate.toISOString() : <any>undefined;
        data["url"] = this.url;
        if (Array.isArray(this.tags)) {
            data["tags"] = [];
            for (let item of this.tags)
                data["tags"].push(item);
        }
        return data;
    }
}

export interface IJob {
    id?: string;
    website?: string | undefined;
    title?: string | undefined;
    company?: string | undefined;
    location?: Location;
    language?: string | undefined;
    description?: string | undefined;
    salary?: Salary;
    posterName?: string | undefined;
    postedDate?: Date;
    url?: string | undefined;
    tags?: string[] | undefined;
}
// ...
export class Location implements ILocation {
    type?: LocationType;
    city?: string | undefined;
    country?: string | undefined;

    constructor(data?: ILocation) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.type = _data["type"];
            this.city = _data["city"];
            this.country = _data["country"];
        }
    }

    static fromJS(data: any): Location {
        data = typeof data === 'object' ? data : {};
        let result = new Location();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["type"] = this.type;
        data["city"] = this.city;
        data["country"] = this.country;
        return data;
    }
}

export interface ILocation {
    type?: LocationType;
    city?: string | undefined;
    country?: string | undefined;
}

export enum LocationType {
    _0 = 0,
    _1 = 1,
    _2 = 2,
}

export class Salary implements ISalary {
    currencyCode?: string | undefined;
    min?: number | undefined;
    max?: number | undefined;
    timeUnit?: SalaryTimeUnit;
    convertedMin?: number | undefined;
    convertedMax?: number | undefined;
    baseCurrencyCode?: string | undefined;
    exchangeRateUpdatedAt?: Date | undefined;

    constructor(data?: ISalary) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.currencyCode = _data["currencyCode"];
            this.min = _data["min"];
            this.max = _data["max"];
            this.timeUnit = _data["timeUnit"];
            this.convertedMin = _data["convertedMin"];
            this.convertedMax = _data["convertedMax"];
            this.baseCurrencyCode = _data["baseCurrencyCode"];
            this.exchangeRateUpdatedAt = _data["exchangeRateUpdatedAt"] ? new Date(_data["exchangeRateUpdatedAt"].toString()) : <any>undefined;
        }
    }

    static fromJS(data: any): Salary {
        data = typeof data === 'object' ? data : {};
        let result = new Salary();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["currencyCode"] = this.currencyCode;
        data["min"] = this.min;
        data["max"] = this.max;
        data["timeUnit"] = this.timeUnit;
        data["convertedMin"] = this.convertedMin;
        data["convertedMax"] = this.convertedMax;
        data["baseCurrencyCode"] = this.baseCurrencyCode;
        data["exchangeRateUpdatedAt"] = this.exchangeRateUpdatedAt ? this.exchangeRateUpdatedAt.toISOString() : <any>undefined;
        return data;
    }
}

export interface ISalary {
    currencyCode?: string | undefined;
    min?: number | undefined;
    max?: number | undefined;
    timeUnit?: SalaryTimeUnit;
    convertedMin?: number | undefined;
    convertedMax?: number | undefined;
    baseCurrencyCode?: string | undefined;
    exchangeRateUpdatedAt?: Date | undefined;
}

export enum SalaryTimeUnit {
    _0 = 0,
    _1 = 1,
    _2 = 2,
    _3 = 3,
    _4 = 4,
}